refactor: rename `/entityOperations` directory to `/api`
