+++ src/api/createEntitiesForTool.ts
@@ -2,7 +2,7 @@ import { isNil, propOr } from "ramda";
 import type { SagaIterator } from "redux-saga";
 import { call, delay, put, select } from "redux-saga/effects";
 
-import { getApiDelayForTool } from "~/entityOperations/apiRequests";
+import { getApiDelayForTool } from "~/api/apiRequests";
 import { entityGroupTransferCompletedCountIncremented } from "~/redux/allEntities/allEntitiesActions";
 import { workspaceIdToLinkedIdSelector } from "~/redux/workspaces/workspacesSelectors";
 import type { EntityGroup, ToolName, ValidEntity } from "~/typeDefs";
