+++ src/redux/clients/clientsSelectors.ts
@@ -1,6 +1,6 @@
 import { isNil, propOr } from "ramda";
 
-import { selectIdToLinkedId } from "~/entityOperations/selectIdToLinkedId";
+import { selectIdToLinkedId } from "~/api/selectIdToLinkedId";
 import { mappingByToolNameSelector } from "~/redux/allEntities/allEntitiesSelectors";
 import { createSelector, type Selector } from "~/redux/reduxTools";
 import { sourceTimeEntryCountByIdFieldSelectorFactory } from "~/redux/timeEntries/timeEntriesSelectors";
