+++ src/redux/clients/sagas/clientsSagas.ts
@@ -2,7 +2,7 @@ import { indexBy, prop } from "ramda";
 import type { SagaIterator } from "redux-saga";
 import { call, put, select } from "redux-saga/effects";
 
-import { linkEntitiesByIdByMapping } from "~/entityOperations/linkEntitiesByIdByMapping";
+import { linkEntitiesByIdByMapping } from "~/api/linkEntitiesByIdByMapping";
 import {
   toolActionSelector,
   toolNameByMappingSelector,
