+++ src/redux/projects/sagas/togglProjectsSagas.ts
@@ -7,10 +7,10 @@ import {
   fetchEmpty,
   fetchObject,
   getApiDelayForTool,
-} from "~/entityOperations/apiRequests";
-import { createEntitiesForTool } from "~/entityOperations/createEntitiesForTool";
-import { deleteEntitiesForTool } from "~/entityOperations/deleteEntitiesForTool";
-import { fetchEntitiesForTool } from "~/entityOperations/fetchEntitiesForTool";
+} from "~/api/apiRequests";
+import { createEntitiesForTool } from "~/api/createEntitiesForTool";
+import { deleteEntitiesForTool } from "~/api/deleteEntitiesForTool";
+import { fetchEntitiesForTool } from "~/api/fetchEntitiesForTool";
 import { clientIdToLinkedIdSelector } from "~/redux/clients/clientsSelectors";
 import { EntityGroup, ToolName, type Project } from "~/typeDefs";
 import { validStringify } from "~/utilities/textTransforms";
