+++ src/redux/workspaces/sagas/togglWorkspacesSagas.ts
@@ -1,7 +1,7 @@
 import type { SagaIterator } from "redux-saga";
 import { call } from "redux-saga/effects";
 
-import { fetchArray } from "~/entityOperations/apiRequests";
+import { fetchArray } from "~/api/apiRequests";
 import { EntityGroup, type Workspace } from "~/typeDefs";
 
 interface TogglWorkspaceSubscription {
