+++ src/utilities/replaceMappingWithToolName.ts
